float -> double 수정

+++ app/src/main/java/com/example/Colorful_Daegu/model/Location.java
@@ -1,16 +1,16 @@
 package com.example.Colorful_Daegu.model;
 
 public class Location {
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
 
     public Location(){}
-    public Location(float latitude, float longitude) {
+    public Location(double latitude, double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
@@ -18,7 +18,7 @@ public class Location {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
